Move ConfigurationError to exceptions module
