+++ checkQC/exceptions.py
@@ -32,5 +32,10 @@ class RunParametersNotFound(CheckQCException):
 class StatsJsonNotFound(CheckQCException):
     pass
 
+
 class QCHandlerNotFound(CheckQCException):
+    pass
+
+
+class ConfigurationError(CheckQCException):
     pass
